ouchhungry_airflow/requirements.txt
# requirements.txt

# Manipulação e análise de dados
pandas>=1.3.0
pandas-gbq>=0.15.0
google-cloud-bigquery>=2.30.0

# Airflow Providers
apache-airflow-providers-google>=6.0.0
apache-airflow-providers-ssh>=2.0.0
apache-airflow-providers-http>=2.0.0

# Dependências adicionais (opcional)
google-auth>=2.0.0
requests>=2.25.0

